Added AbiStructSha3KeccackHashByteArrayConvertor and AbiStructSha3KeccackMerkleTree abstractions.
