+++ src/Nethereum.ABI/ByteArrayConvertors/AbiStructSha3KeccackHashByteArrayConvertor.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Nethereum.Util;
+using Nethereum.Util.ByteArrayConvertors;
+using Nethereum.Util.HashProviders;
+
+
+namespace Nethereum.ABI.ByteArrayConvertors
+{
+    internal class AbiStructSha3KeccackHashByteArrayConvertor<T> : IByteArrayConvertor<T>
+    {
+        private readonly ABIEncode _abiEncode;
+        public AbiStructSha3KeccackHashByteArrayConvertor()
+        {
+            _abiEncode = new ABIEncode();
+        }
+        public byte[] ConvertToByteArray(T data)
+        {
+            var encoded = _abiEncode.GetABIParamsEncoded(data);
+            return Sha3Keccack.Current.CalculateHash(encoded);
+
+        }
+    }
+}
